move public api to its own file
