+++ client/src/leap/soledad/client/sync.py
@@ -14,8 +14,6 @@
 #
 # You should have received a copy of the GNU General Public License
 # along with this program. If not, see <http://www.gnu.org/licenses/>.
-
-
 """
 Soledad synchronization utilities.
 
@@ -27,8 +25,6 @@ Extend u1db Synchronizer with the ability to:
 
     * Be interrupted and recovered.
 """
-
-
 import logging
 import traceback
 from threading import Lock
